+++ src/EntryTypes/Hooks/BeforeHydratingHook.php
@@ -2,32 +2,36 @@
 
 namespace CbtechLtd\Fastlane\EntryTypes\Hooks;
 
+use CbtechLtd\Fastlane\Support\Contracts\EntryInstance;
 use CbtechLtd\Fastlane\Support\Contracts\EntryType;
 use Illuminate\Database\Eloquent\Model;
 
 class BeforeHydratingHook
 {
-    protected EntryType $entryType;
-    protected Model $model;
+    protected EntryInstance $entryInstance;
     protected array $fields;
     public array $data;
 
-    public function __construct(EntryType $entryType, Model $model, array $fields, array $data)
+    public function __construct(EntryInstance $entryInstance, array $fields, array $data)
     {
-        $this->entryType = $entryType;
-        $this->model = $model;
+        $this->entryInstance = $entryInstance;
         $this->fields = $fields;
         $this->data = $data;
     }
 
+    public function entryInstance(): EntryInstance
+    {
+        return $this->entryInstance;
+    }
+
     public function entryType(): EntryType
     {
-        return $this->entryType;
+        return $this->entryInstance->type();
     }
 
     public function model(): Model
     {
-        return $this->model;
+        return $this->entryInstance->model();
     }
 
     public function fields(): array