+++ src/Support/Contracts/EntryType.php
@@ -2,16 +2,13 @@
 
 namespace CbtechLtd\Fastlane\Support\Contracts;
 
-use CbtechLtd\Fastlane\Support\Schema\Contracts\EntrySchema;
-use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\Request;
+use Illuminate\Support\Collection;
 
 interface EntryType extends Hookable
 {
-    public function new(?Model $model): EntryType;
-
-    public function resolve(array $requestData): EntryType;
+    public function newInstance(?Model $model): EntryInstance;
 
     public function identifier(): string;
 
@@ -23,8 +20,6 @@ interface EntryType extends Hookable
 
     public function model(): string;
 
-    public function modelInstance(): Model;
-
     public function newModelInstance(): Model;
 
     public function apiResource(): string;
@@ -35,21 +30,17 @@ interface EntryType extends Hookable
 
     public function fields(): array;
 
-    public function schema(): EntrySchema;
-
     public function isVisibleOnMenu(): bool;
 
     public function getItems(): Collection;
 
-    public function findItem(string $hashid): Model;
+    public function findItem(string $hashid): EntryInstance;
 
-    public function store(Request $request): Model;
+    public function store(Request $request): EntryInstance;
 
-    public function update(Request $request, string $hashid): Model;
+    public function update(Request $request, string $hashid): EntryInstance;
 
-    public function delete(string $hashid): Model;
+    public function delete(string $hashid): EntryInstance;
 
     public function install(): void;
-
-    public function makeModelTitle(Model $model): string;
 }