+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to `fastlane` will be documented in this file
 
 ## [Unreleased]
 
+## 0.4.0 [2020-08-23]
+
+### Changed
+- How entry types and fields are implemented.
+- Schema usage by entry types / fields.
+- Small tweaks on style.
+
+---
+
 ## 0.3.0 [2020-08-20]
 
 ### Added